Finalizando desafio

+++ src/shared/infra/typeorm/migrations/1603850122652-CreateProducts.ts
@@ -26,7 +26,7 @@ export class CreateProducts1603850122652 implements MigrationInterface {
                 },
                 {
                     name: "quantity",
-                    type: "bigint",
+                    type: "int",
                 },
                 {
                     name: "created_at",
